fix: output invalid packages with invalid semver
